requirements.txt
opencv-python
numpy
pandas
tensorflow-gpu
tqdm
scikit-learn
GPUtil
pyrex-core
psuti